delete hand.py, rename Hand to ShareCards, and move it to cards.py
